page initialization and profile component for header
